lttng-tools ptest: add missing dependencies
Multiple tests are failing due to missing dependencies on a bare
core-image-minimal build with only lttng-tools ptest present.

"getconf LONG_BIT" is used to get the bitness of the host to run the
correct consumerd. Depend on glibc-utils.

perl-module-overloading is needed to validate trace. babelstats.pl

"head -c" is used to generate a random string. It is easier to depends
on coreutils to get a feature complete "head" than to enable to "fancy
head" feature of busybox. See utils.sh randstring.

"taskset" is used in a couple of scripts to ensure that events generated
finish in the same buffer. Depend on util-linux.

(From OE-Core rev: 55ed568c2d87447ea513798814ecd9f65f5d4d15)

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>